requirements.txt
fastapi==0.65.2
gunicorn==20.0.4
pydantic==1.6.2
uvicorn==0.11.7
tensorflow-cpu==2.9.0